Move reports to a component.
