+++ app/components/reports/__init__.py
@@ -1,0 +1,8 @@
+from . import crud, models, routes, schemas
+
+__all__ = [
+    "crud",
+    "models",
+    "routes",
+    "schemas",
+]
