+++ app/components/reports/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, ForeignKey
 from sqlalchemy.orm import relationship
 
 from app.db.base_class import Base
@@ -10,3 +10,10 @@ class Report(Base):
 
     verbose_name = Column(String, nullable=False, unique=True)
     options = relationship("ReportOption")
+
+
+class ReportOption(Base):
+
+    id = Column(Integer, primary_key=True, index=True)
+    verbose_name = Column(String)
+    report_id = Column(Integer, ForeignKey("report.id"))
