+++ app/db/base.py
@@ -2,13 +2,12 @@ from app.components.geospatial.models import GeospatialIndex
 from app.components.locations.models import Location
 from app.components.oauth.models import OauthScope, OauthRole, association_table
 from app.components.organizations.models import Organization
+from app.components.reports.models import Report, ReportOption
 from app.components.users.models import User
 from app.components.zones.models import Zone
 from app.db.base_class import Base
 from app.models.changelog import ChangeLog
 from app.models.guest_user import GuestUser
-from app.models.reportoptions import ReportOption
-from app.models.reports import Report
 from app.models.sessionhistory import SessionHistory
 
 __all__ = [
