+++ app/models/__init__.py
@@ -1,4 +1,2 @@
 from .changelog import ChangeLog
-from .reportoptions import ReportOption
-from .reports import Report
 from .sessionhistory import SessionHistory
